requirements.txt
langchain_chroma
langchain_huggingface
langchain_openai
langchain_classic
rich
python-dotenv
sentence-transformers